requirements.txt
langchain
llama-index
numpy
openai
pandas
pydantic
python-dotenv
scikit-learn
scipy
sentence-transformers
sentencepiece
tokenizers
SQLAlchemy